(none)
git-svn-id: https://svn.code.sf.net/p/khc/svn/trunk@228 ab1598f5-0ef3-491f-8894-8c1f2390833e

+++ core/kh_lib/kh_dictio.pm
@@ -106,7 +106,7 @@ sub mark{
 	my $dist   = $::project_obj->file_m_target;
 
 	unless (eval (@{$self->words_mk})){
-		unlink($dist) or die;
+		unlink($dist) or die if -e $dist;
 		use File::Copy;
 		copy("$source","$dist") or die;
 		return undef;
